cleanup, rewrite Placeholder - initial implementation
